More LCD definition around and some fixes

+++ TeaTimeAdvance/Device/IO/LCD/LCDStatus.cs
@@ -3,6 +3,6 @@
     public record struct LCDStatus
     {
         public LCDStatusFlags Flags;
-        public ushort VerticalCount;
+        public byte VerticalCount;
     }
 }
